Prover: improve numerical stability by dynamically set precision for free points (GGB-1371)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@52336 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/cas/giac/CASgiac.java
@@ -738,15 +738,14 @@ public abstract class CASgiac implements CASGenericInterface {
 	 */
 	@Override
 	public String createEliminateScript(String polys, String elimVars,
-			boolean oneCurve, Integer precision) {
+			boolean oneCurve, Long precision) {
 		if (!oneCurve) {
 			return CustomFunctions.PRIM_POLY + "(" + CustomFunctions.ELIMINATE2
 					+ "([" + polys
 					+ "],revlist([" + elimVars + "])))";
 		}
 
-		String PRECISION = Integer.toString(precision);
-		// PRECISION = "10000";
+		String PRECISION = Long.toString(precision);
 		Log.debug("PRECISION = " + PRECISION);
 		String retval;
 		/*
